// web/src/types/item.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type Item = { id: number, link: string, title: string | null, description: string | null, author: string | null, published: string | null, source_link: string | null, image: string | null, favorite: boolean, done: boolean, created_at: string, updated_at: string, source_id: bigint | null, };